// src/openapi/src/models/Platform.ts
/* tslint:disable */
/* eslint-disable */
/**
 * MTSynergy Platform BFF API
 * Backend For Frontend API for Web and Mobile clients
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Social media platform
 * @export
 */
export const Platform = {
    Twitter: 'twitter',
    Tiktok: 'tiktok',
    Facebook: 'facebook',
    Instagram: 'instagram',
    Linkedin: 'linkedin',
    Youtube: 'youtube'
} as const;
export type Platform = typeof Platform[keyof typeof Platform];


export function instanceOfPlatform(value: any): boolean {
    for (const key in Platform) {
        if (Object.prototype.hasOwnProperty.call(Platform, key)) {
            if (Platform[key as keyof typeof Platform] === value) {
                return true;
            }
        }
    }
    return false;
}

export function PlatformFromJSON(json: any): Platform {
    return PlatformFromJSONTyped(json, false);
}

export function PlatformFromJSONTyped(json: any, ignoreDiscriminator: boolean): Platform {
    return json as Platform;
}

export function PlatformToJSON(value?: Platform | null): any {
    return value as any;
}

export function PlatformToJSONTyped(value: any, ignoreDiscriminator: boolean): Platform {
    return value as Platform;
}