// src/openapi/src/models/CreateDraftRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * MTSynergy Platform BFF API
 * Backend For Frontend API for Web and Mobile clients
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Platform } from './Platform';
import {
    PlatformFromJSON,
    PlatformFromJSONTyped,
    PlatformToJSON,
    PlatformToJSONTyped,
} from './Platform';

/**
 * 
 * @export
 * @interface CreateDraftRequest
 */
export interface CreateDraftRequest {
    /**
     * Post caption/text
     * @type {string}
     * @memberof CreateDraftRequest
     */
    caption: string;
    /**
     * Platforms to publish to
     * @type {Array<Platform>}
     * @memberof CreateDraftRequest
     */
    platforms: Array<Platform>;
    /**
     * ISO8601 scheduled publish time
     * @type {Date}
     * @memberof CreateDraftRequest
     */
    scheduledAt?: Date | null;
    /**
     * IDs of media files to attach
     * @type {Array<string>}
     * @memberof CreateDraftRequest
     */
    mediaIds?: Array<string> | null;
}

/**
 * Check if a given object implements the CreateDraftRequest interface.
 */
export function instanceOfCreateDraftRequest(value: object): value is CreateDraftRequest {
    if (!('caption' in value) || value['caption'] === undefined) return false;
    if (!('platforms' in value) || value['platforms'] === undefined) return false;
    return true;
}

export function CreateDraftRequestFromJSON(json: any): CreateDraftRequest {
    return CreateDraftRequestFromJSONTyped(json, false);
}

export function CreateDraftRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateDraftRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'caption': json['caption'],
        'platforms': ((json['platforms'] as Array<any>).map(PlatformFromJSON)),
        'scheduledAt': json['scheduledAt'] == null ? undefined : (new Date(json['scheduledAt'])),
        'mediaIds': json['mediaIds'] == null ? undefined : json['mediaIds'],
    };
}

export function CreateDraftRequestToJSON(json: any): CreateDraftRequest {
    return CreateDraftRequestToJSONTyped(json, false);
}

export function CreateDraftRequestToJSONTyped(value?: CreateDraftRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'caption': value['caption'],
        'platforms': ((value['platforms'] as Array<any>).map(PlatformToJSON)),
        'scheduledAt': value['scheduledAt'] == null ? value['scheduledAt'] : value['scheduledAt'].toISOString(),
        'mediaIds': value['mediaIds'],
    };
}