// src/openapi/src/models/CreateDraftResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * MTSynergy Platform BFF API
 * Backend For Frontend API for Web and Mobile clients
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateDraftResponse
 */
export interface CreateDraftResponse {
    /**
     * Draft ID
     * @type {string}
     * @memberof CreateDraftResponse
     */
    id: string;
    /**
     * Draft status
     * @type {string}
     * @memberof CreateDraftResponse
     */
    status: CreateDraftResponseStatusEnum;
    /**
     * Creation timestamp
     * @type {Date}
     * @memberof CreateDraftResponse
     */
    createdAt: Date;
    /**
     * Last update timestamp
     * @type {Date}
     * @memberof CreateDraftResponse
     */
    updatedAt: Date;
}


/**
 * @export
 */
export const CreateDraftResponseStatusEnum = {
    Draft: 'DRAFT',
    PendingApproval: 'PENDING_APPROVAL',
    Approved: 'APPROVED'
} as const;
export type CreateDraftResponseStatusEnum = typeof CreateDraftResponseStatusEnum[keyof typeof CreateDraftResponseStatusEnum];


/**
 * Check if a given object implements the CreateDraftResponse interface.
 */
export function instanceOfCreateDraftResponse(value: object): value is CreateDraftResponse {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('status' in value) || value['status'] === undefined) return false;
    if (!('createdAt' in value) || value['createdAt'] === undefined) return false;
    if (!('updatedAt' in value) || value['updatedAt'] === undefined) return false;
    return true;
}

export function CreateDraftResponseFromJSON(json: any): CreateDraftResponse {
    return CreateDraftResponseFromJSONTyped(json, false);
}

export function CreateDraftResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateDraftResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'status': json['status'],
        'createdAt': (new Date(json['createdAt'])),
        'updatedAt': (new Date(json['updatedAt'])),
    };
}

export function CreateDraftResponseToJSON(json: any): CreateDraftResponse {
    return CreateDraftResponseToJSONTyped(json, false);
}

export function CreateDraftResponseToJSONTyped(value?: CreateDraftResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'status': value['status'],
        'createdAt': value['createdAt'].toISOString(),
        'updatedAt': value['updatedAt'].toISOString(),
    };
}