// src/openapi/src/models/InboxItem.ts
export interface InboxItem {
    /**
     * 
     * @type {string}
     * @memberof InboxItem
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof InboxItem
     */
    workspaceId: string;
    /**
     * 
     * @type {Platform}
     * @memberof InboxItem
     */
    platform: Platform;
    /**
     * 
     * @type {string}
     * @memberof InboxItem
     */
    platformItemId: string;
    /**
     * 
     * @type {string}
     * @memberof InboxItem
     */
    author: string;
    /**
     * 
     * @type {string}
     * @memberof InboxItem
     */
    content: string;
    /**
     * 
     * @type {string}
     * @memberof InboxItem
     */
    contentType: InboxItemContentTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof InboxItem
     */
    status: InboxItemStatusEnum;
    /**
     * 
     * @type {Date}
     * @memberof InboxItem
     */
    createdAt: Date;
}


/**
 * @export
 */
export const InboxItemContentTypeEnum = {
    Comment: 'COMMENT',
    Message: 'MESSAGE',
    Mention: 'MENTION'
} as const;
export type InboxItemContentTypeEnum = typeof InboxItemContentTypeEnum[keyof typeof InboxItemContentTypeEnum];

/**
 * @export
 */
export const InboxItemStatusEnum = {
    New: 'NEW',
    Assigned: 'ASSIGNED',
    Resolved: 'RESOLVED',
    Spam: 'SPAM'
} as const;
export type InboxItemStatusEnum = typeof InboxItemStatusEnum[keyof typeof InboxItemStatusEnum];