// src/openapi/src/models/ListInboxResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * MTSynergy Platform BFF API
 * Backend For Frontend API for Web and Mobile clients
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { InboxItem } from './InboxItem';
import {
    InboxItemFromJSON,
    InboxItemFromJSONTyped,
    InboxItemToJSON,
    InboxItemToJSONTyped,
} from './InboxItem';
import type { ListInboxResponsePagination } from './ListInboxResponsePagination';
import {
    ListInboxResponsePaginationFromJSON,
    ListInboxResponsePaginationFromJSONTyped,
    ListInboxResponsePaginationToJSON,
    ListInboxResponsePaginationToJSONTyped,
} from './ListInboxResponsePagination';

/**
 * 
 * @export
 * @interface ListInboxResponse
 */
export interface ListInboxResponse {
    /**
     * 
     * @type {Array<InboxItem>}
     * @memberof ListInboxResponse
     */
    items: Array<InboxItem>;
    /**
     * 
     * @type {ListInboxResponsePagination}
     * @memberof ListInboxResponse
     */
    pagination: ListInboxResponsePagination;
}

/**
 * Check if a given object implements the ListInboxResponse interface.
 */
export function instanceOfListInboxResponse(value: object): value is ListInboxResponse {
    if (!('items' in value) || value['items'] === undefined) return false;
    if (!('pagination' in value) || value['pagination'] === undefined) return false;
    return true;
}

export function ListInboxResponseFromJSON(json: any): ListInboxResponse {
    return ListInboxResponseFromJSONTyped(json, false);
}

export function ListInboxResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): ListInboxResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'items': ((json['items'] as Array<any>).map(InboxItemFromJSON)),
        'pagination': ListInboxResponsePaginationFromJSON(json['pagination']),
    };
}

export function ListInboxResponseToJSON(json: any): ListInboxResponse {
    return ListInboxResponseToJSONTyped(json, false);
}

export function ListInboxResponseToJSONTyped(value?: ListInboxResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'items': ((value['items'] as Array<any>).map(InboxItemToJSON)),
        'pagination': ListInboxResponsePaginationToJSON(value['pagination']),
    };
}