// src/openapi/src/models/ListInboxResponsePagination.ts
/* tslint:disable */
/* eslint-disable */
/**
 * MTSynergy Platform BFF API
 * Backend For Frontend API for Web and Mobile clients
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ListInboxResponsePagination
 */
export interface ListInboxResponsePagination {
    /**
     * 
     * @type {number}
     * @memberof ListInboxResponsePagination
     */
    page: number;
    /**
     * 
     * @type {number}
     * @memberof ListInboxResponsePagination
     */
    pageSize: number;
    /**
     * 
     * @type {number}
     * @memberof ListInboxResponsePagination
     */
    total: number;
}

/**
 * Check if a given object implements the ListInboxResponsePagination interface.
 */
export function instanceOfListInboxResponsePagination(value: object): value is ListInboxResponsePagination {
    if (!('page' in value) || value['page'] === undefined) return false;
    if (!('pageSize' in value) || value['pageSize'] === undefined) return false;
    if (!('total' in value) || value['total'] === undefined) return false;
    return true;
}

export function ListInboxResponsePaginationFromJSON(json: any): ListInboxResponsePagination {
    return ListInboxResponsePaginationFromJSONTyped(json, false);
}

export function ListInboxResponsePaginationFromJSONTyped(json: any, ignoreDiscriminator: boolean): ListInboxResponsePagination {
    if (json == null) {
        return json;
    }
    return {
        
        'page': json['page'],
        'pageSize': json['pageSize'],
        'total': json['total'],
    };
}

export function ListInboxResponsePaginationToJSON(json: any): ListInboxResponsePagination {
    return ListInboxResponsePaginationToJSONTyped(json, false);
}

export function ListInboxResponsePaginationToJSONTyped(value?: ListInboxResponsePagination | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'page': value['page'],
        'pageSize': value['pageSize'],
        'total': value['total'],
    };
}